pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "A ReAct (Reason and Act) AI agent that handles multiple tool calls"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.11.4",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
    "yfinance>=0.2.54",
]
